fix #40800 (cannot disable memory_limit with -1)

+++ Zend/zend_alloc.c
@@ -2133,7 +2133,7 @@ ZEND_API zstr zend_zstrndup(int type, const zstr s, uint length)
 	}
 }
 
-ZEND_API int zend_set_memory_limit(unsigned int memory_limit)
+ZEND_API int zend_set_memory_limit(size_t memory_limit)
 {
 	TSRMLS_FETCH();
